requirements.txt
pyspark==3.5.0
cassandra-driver==3.28.0
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0 